+++ metadata.rb
@@ -1,8 +1,8 @@
-name 'example'
+name 'skeleton'
 maintainer 'Numergy'
 maintainer_email 'cd@numergy.com'
 license 'Numergy'
-description 'Installs/Configures example'
+description 'Installs/Configures skeleton'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.1.0'