+++ spec/recipes/default_spec.rb
@@ -4,7 +4,7 @@
 #
 require_relative '../spec_helper'
 
-describe 'example::default' do
+describe 'skeleton::default' do
   subject { ChefSpec::Runner.new.converge(described_recipe) }
 
   # Write full examples using the `expect` syntax