+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-example CHANGELOG
+skeleton CHANGELOG
 =================
 
-This file is used to list changes made in each version of the example cookbook.
+This file is used to list changes made in each version of the skeleton cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of example
+- [your_name] - Initial release of skeleton
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.